backend/pyproject.toml
[tool.poetry]
name = "chatweb"
version = "0.1.0"
description = ""
authors = ["David Gootman <gootman@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.103.2"
uvicorn = {extras = ["standard"], version = "^0.23.2"}
loguru = "^0.7.2"
aiohttp = "^3.8.5"
python-slugify = "^8.0.1"

[tool.poetry.group.dev.dependencies]
black = "^23.9.1"
isort = "^5.12.0"
flake8 = "^6.1.0"
pylint = "^3.0.0"

[tool.pylint."messages control"]
disable = ["missing-class-docstring", "missing-function-docstring", "missing-module-docstring"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
